requirements.txt
asn1crypto==0.24.0
boto3==1.9.161
botocore==1.12.161
certifi==2019.3.9
cffi==1.12.3
cryptography==2.7
Django==2.2.2
django-cors-headers==3.0.2
django-rest-knox==4.1.0
django-storages==1.7.1
django-webpack-loader==0.6.0
djangorestframework==3.9.4
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.4
pathlib==1.0.1
Pillow==6.0.0
pipenv==2018.11.26
psycopg2==2.8.2
pycparser==2.19
python-dateutil==2.8.0
pytz==2019.1
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
virtualenv==16.6.0
virtualenv-clone==0.5.3
